feat: create a router and move home folder to features folder
